Updates to utils seperating funtions to different files
